requirements.txt
asgiref==3.5.2
autopep8==1.6.0
Django==4.0.4
django-cors-headers==3.12.0
django-debug-toolbar==3.4.0
djangorestframework==3.13.1
pycodestyle==2.8.0
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
